+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 * Add 'User: Can override patron blocks' permission. Refs UIU-2025.
 * Add 'User: Can override item blocks' permission. Refs UIU-2019.
 * Add EXPORT option to User Information. Refs UIU-1957.
+* Add EXPORT option to fees/fines Details. Refs UIU-1956.
 
 ## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v5.0.0...v5.0.1)